Majorly bulldozed the applicator_geometry stuff :D. Let's see what happens.
