+++ brachyutils/planning/plan_utils.py
@@ -572,150 +572,7 @@ class BrachyPlan:
                 ["body"], ROIContour, strict_name_match=False
             ).get("body", None)
 
-    def load_applicator_list(
-        self,
-        applicator_list_pth: Union[list, Path, str],
-        format: Literal["WebApp", "RapidBrachy"] = "WebApp",
-    ):
-        r"""
-        ### Purpose:
-        - To load the applicator list from a json file containing the applicator geometry.
-
-        ### Inputs:
-        - applicator_list_pth:str := path to the json file containing the applicator list with N applicators.
-        The items inside this list have the attributes bellow. If any left empty, the default value will be used.
-        these attributes could be changed later using the setter functions.
-
-        if the format is WebApp, the attributes are:
-            - "path": path to the applicator geometry file (.stl or .json).
-            - "material": material of the applicator (str).
-            - "density": density of the applicator (str).
-            - "origin": origin of the applicator ([x,y,z]).
-            - "rotation": rotation of the applicator ([w,x,y,z]).
-            - "rotation_origin": origin of the rotation ([x,y,z]).
-            - "coordinates": coordinates of the applicator ([x,y,z]).
-
-        if the format is RapidBrachy, the attributes are:
-            - "densities": list of densities of the applicator.
-            - "filenames": list of filenames of the applicator.
-            - "materials": list of materials of the applicator.
-            - "points": list of points (x,y,z,x,y,z) describing the first and last dwell positions
-            on the applicator in the frame of the applicator.
-            - "wRot": list of wRot of the applicator.
-            - "x": list of x of the applicator.
-            - "xRoti": list of xRot of the applicator i in [1, N].
-            - "y": list of y of the applicator.
-            - "yRoti": list of yRot of the applicator i in [1, N].
-            - "z": list of z of the applicator.
-            - "zRoti": list of zRot of the applicator i in [1, N].
-        - format:str := the format of the applicator geometry file. options are "RapidBrachy" or "WebApp"
-
-        ### Outputs:
-            - None := will update the BrachyPlan.applicator_list attribute
-        """
-        if isinstance(applicator_list_pth, Path) or isinstance(
-            applicator_list_pth, str
-        ):
-            with open(applicator_list_pth, "r") as json_file:
-                applicator_list = json.load(json_file)
-        if format == "RapidBrachy":
-            num_applicators = len(applicator_list["densities"])
-
-            for i in range(num_applicators):
-
-                j = i + 1 if i >= 1 else ""
-
-                rotation = np.array(
-                    [
-                        (
-                            applicator_list[f"wRot{j}"]
-                            if f"wRot{j}" in applicator_list
-                            else 0
-                        ),
-                        (
-                            applicator_list[f"xRot{j}"]
-                            if f"xRot{j}" in applicator_list
-                            else 0
-                        ),
-                        (
-                            applicator_list[f"yRot{j}"]
-                            if f"yRot{j}" in applicator_list
-                            else 0
-                        ),
-                        (
-                            applicator_list[f"zRot{j}"]
-                            if f"zRot{j}" in applicator_list
-                            else 0
-                        ),
-                    ]
-                )
-
-                applicator_obj = BrachyApplicator(
-                    pth_input_file=applicator_list["filenames"][i],
-                    material=applicator_list["materials"][i],
-                    density=applicator_list["densities"][i],
-                    origin=self.patient_origin,
-                    rotation=rotation,
-                    rotation_origin=np.array(
-                        [
-                            applicator_list["x"],
-                            applicator_list["y"],
-                            applicator_list["z"],
-                        ]
-                    ),
-                    coordinates=np.array(
-                        [
-                            applicator_list["x"],
-                            applicator_list["y"],
-                            applicator_list["z"],
-                        ]
-                    ),
-                    # for now RapidBrachy exports only one catheter trajectory.
-                    # in future, more catheter trajectories may be possible.
-                    # use i instead of 0 to get the ith catheter trajectory.
-                    catheter_trajectory=np.array(
-                        [
-                            applicator_list["points"][0][0:3],
-                            applicator_list["points"][0][3:6],
-                        ]
-                    ),
-                )
-
-                self.applicator_list.append(applicator_obj)
-
-        elif format == "WebApp":
-            for applicator in applicator_list:
-
-                applicator_obj = BrachyApplicator(
-                    pth_input_file=applicator["path"] if "path" in applicator else None,
-                    material=(
-                        applicator["material"] if "material" in applicator else None
-                    ),
-                    density=applicator["density"] if "density" in applicator else None,
-                    origin=applicator["origin"] if "origin" in applicator else None,
-                    rotation=(
-                        applicator["rotation"] if "rotation" in applicator else None
-                    ),
-                    rotation_origin=(
-                        applicator["rotation_origin"]
-                        if "rotation_origin" in applicator
-                        else None
-                    ),
-                    coordinates=(
-                        applicator["coordinates"]
-                        if "coordinates" in applicator
-                        else None
-                    ),
-                    catheter_trajectory=(
-                        applicator["catheter_trajectory"]
-                        if "catheter_trajectory" in applicator
-                        else None
-                    ),
-                )
-                self.applicator_list.append(applicator_obj)
-        else:
-            raise ValueError("format should be either 'RapidBrachy' or 'WebApp'")
-
+   
     def get_dvh_metrics(
         self,
         combined_dose: BrachyDose=None,
@@ -860,7 +717,7 @@ class BrachyPlan:
             - Plan files (.plan files for each dwell position)
             - Macro files (.mac simulation files)
             - Egsphant phantom file (ct.egsphant)
-            - Applicator geometry files (applicator_geometry.json and .mac files)
+            - Applicator stl files
             - Structure set file (structure_set.json)
 
         ### Dependencies:
@@ -870,7 +727,6 @@ class BrachyPlan:
         - export_plan_files()
         - export_mac_files()
         - _export_egsphant()
-        - _export_applicator_geometry()
         - _export_structure_set()
 
         ### Example of content_to_export dict:
@@ -933,7 +789,8 @@ class BrachyPlan:
 
 
         if content_to_export.applicator_geometry:
-            self._export_applicator_geometry(str(content_to_export.dir_export))
+            for applicator in self.applicator_list:
+                applicator.to_stl(dir_export / f"{applicator.name}.stl")
 
         if content_to_export.structure_set:
             self._export_structure_set(
@@ -1171,79 +1028,6 @@ class BrachyPlan:
 
         print("Egsphant file was exported successfully")
 
-    def _export_applicator_geometry(
-        self, dir_export: str, export_format: str = "RapidBrachy"
-    ):
-        r"""
-        ### Purpose:
-        - To export the applicator geometries either in the RapidBrachy Format (mac files and single json file)
-        or in webapp format (json file).
-
-        ### Inputs:
-        - dir_export := path to the directory where the export happens
-        - format := the format of the applicator geometry file. options are "RapidBrachy" or "WebApp"
-
-        ### Outputs:
-        - None := will export the applicator geometries into the specified export directory.
-
-        ### Dependencies:
-        - None
-        """
-        if export_format == "RapidBrachy":
-
-            # initialize the fields of the json file:
-            out_json = {
-                "densities": [],
-                "filenames": [],
-                "materials": [],
-                "points": [],
-                "wRot": 0,
-                "x": 0,
-                "xRot": 0,
-                "y": 0,
-                "yRot": 0,
-                "z": 0,
-                "zRot": 0,
-            }
-            counter = 0
-            for applicator in self.applicator_list:
-
-                out_json["densities"].append(applicator.density)
-                out_json["filenames"].append(applicator.path)
-                out_json["materials"].append(applicator.material)
-                out_json["points"].append(
-                    applicator.catheter_trajectory.flatten().tolist()
-                )
-
-                subscript = counter + 1 if counter >= 1 else ""
-                out_json[f"wRot{subscript}"] = float(applicator.rotation[0])
-                out_json[f"xRot{subscript}"] = float(applicator.rotation[1])
-                out_json[f"yRot{subscript}"] = float(applicator.rotation[2])
-                out_json[f"zRot{subscript}"] = float(applicator.rotation[3])
-
-                out_json["x"] = float(applicator.coordinates[0])
-                out_json["y"] = float(applicator.coordinates[1])
-                out_json["z"] = float(applicator.coordinates[2])
-                counter += 1
-
-            with open(dir_export + "/applicator_geometry.json", "w") as file:
-                json.dump(out_json, file, indent=4)
-
-        elif export_format == "WebApp":
-            out_json = [
-                applicator.to_dict(format) for applicator in self.applicator_list
-            ]
-            with open(dir_export + "/applicator_geometry.json", "w") as file:
-                json.dump(out_json, file, indent=4)
-
-        else:
-            raise ValueError("format should be either 'RapidBrachy' or 'WebApp'")
-
-        # export the transformed stl files for each applicator
-        for applicator in self.applicator_list:
-            applicator.to_stl(os.path.join(dir_export, f"{applicator.name}.stl"))
-        print("applicator geometry file was exported successfully")
-
     def _export_structure_set(
         self,
         dir_export: str,
@@ -1272,7 +1056,7 @@ class BrachyPlan:
         raise NotImplementedError("now that you are here, finish this function thank you!") #no thanks, Jonathan.
         structure_set = []
         for structure in self.structure_list:
-            structure_set.append(structure.to_dict(export_format))
+            structure_set.append(structure.to_dict())
 
             if materials_table is not None:
                 from brachyutils.geometry.egsphant_utils import _load_material_dict
